requirements.txt
# ============================================================
#  MindForge — pip install -r requirements.txt
#  Run this FIRST before anything else
# ============================================================

# ── Core ML / Data ───────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
shap>=0.44.0
imbalanced-learn>=0.11.0

# ── Deep Learning (PyTorch — install CPU or CUDA separately) ─
# torch>=2.2.0                    # installed separately (see README)
transformers>=4.40.0
datasets>=2.18.0
tokenizers>=0.19.0
accelerate>=0.29.0
peft>=0.10.0                      # LoRA / QLoRA
trl>=0.8.6                        # SFTTrainer (supervised fine-tuning)
bitsandbytes>=0.43.0              # 4-bit / 8-bit quantization

# ── Unsloth (2x faster fine-tuning) ──────────────────────────
# Install with: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# OR for local CUDA: pip install unsloth

# ── RAG Stack ────────────────────────────────────────────────
langchain>=0.1.20
langchain-community>=0.0.38
langchain-huggingface>=0.0.3
chromadb>=0.4.24
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0                  # swap for faiss-gpu if you have CUDA

# ── Experiment Tracking ───────────────────────────────────────
wandb>=0.16.0
mlflow>=2.12.0

# ── API & App ────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.33.0
gradio>=4.26.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0

# ── Utilities ────────────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.0
typer>=0.12.0
joblib>=1.4.0
plotly>=5.21.0
matplotlib>=3.8.0
seaborn>=0.13.0
nltk>=3.8.1
regex>=2024.4.28
